requirements.txt
Keras==2.3.1
tqdm==4.56.0
tensorflow_gpu==2.4.1
opencv_python==4.1.2.30
efficientnet==1.1.1
numpy==1.19.5
Pillow==8.2.0
tensorflow==2.5.0
